Add files via upload
It is a HTML, CSS and JavaScript based website. Beats of 7 different drums (a drum kit) are given and the user can play them simply by pressing the buttons as mentioned in the images on the webpage.
Udemy course has lead me to complete this beautiful yet interesting project as the beginner for Web Developer.
All the required files are attached in this repository, I hope it becomes helpful.